Separate demo from component build
