+++ demo/babel.config.js
@@ -14,9 +14,4 @@ module.exports = {
     '@babel/plugin-transform-react-inline-elements',
     'babel-plugin-macros',
   ],
-  "env": {
-    "production": {
-      "presets": ["minify"],
-    }
-  }
 };
