+++ demo/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "react-autocomplete-demo",
+  "version": "1.0.4",
+  "scripts": {
+    "clean": "rimraf dist",
+    "build": "babel ../src --out-dir dist/src --copy-files && babel ./src --out-dir dist/demo/app && webpack",
+    "start": "node ./dist/demo/app/server.js",
+    "lint:js": "eslint src/**/*.js && eslint src/**/*.jsx",
+    "lint:css": "stylelint \"./src/**/*.css\"",
+    "lint:fix": "eslint --fix src/**/*.js && eslint --fix src/**/*.jsx",
+    "lint": "npm run lint:js && npm run lint:css"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.6.3",
+    "express": "^4.17.1",
+    "fetch-everywhere": "^1.0.5",
+    "prop-types": "../node_modules/prop-types",
+    "react": "../node_modules/react",
+    "react-dom": "../node_modules/react-dom"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.6.4",
+    "@babel/core": "^7.6.4",
+    "@babel/node": "^7.6.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
+    "@babel/plugin-syntax-optional-chaining": "^7.2.0",
+    "@babel/plugin-transform-react-inline-elements": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.6.3",
+    "@babel/preset-react": "^7.6.3",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-css-modules-transform": "^1.6.2",
+    "babel-plugin-macros": "^2.6.1",
+    "css-loader": "^3.2.0",
+    "dotenv": "^8.0.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
+    "eslint": "^6.6.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.16.0",
+    "eslint-plugin-standard": "^4.0.1",
+    "husky": "^3.0.8",
+    "jest": "24.9.0",
+    "jest-enzyme": "^7.1.1",
+    "jsdom": "^15.1.1",
+    "mini-css-extract-plugin": "^0.8.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss-cssnext": "^3.1.0",
+    "postcss-import": "^12.0.1",
+    "postcss-loader": "^3.0.0",
+    "react-router-dom": "^5.0.1",
+    "react-test-renderer": "^16.8.0",
+    "rimraf": "^3.0.0",
+    "style-loader": "^1.0.0",
+    "stylelint": "^11.1.1",
+    "stylelint-config-recommended": "^3.0.0",
+    "terser-webpack-plugin": "^2.1.3",
+    "webpack": "^4.41.1",
+    "webpack-cli": "^3.3.9"
+  }
+}
