+++ demo/src/SimpleExample/SimpleExample.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useCallback } from 'react';
-import ReactAutocomplete from '../../src';
+import ReactAutocomplete from '../../../src';
 import { AUTOCOMPLETE_OPTIONS } from '../constants';
 
 const SIMPLE_AUTOCOMPLETE_OPTIONS = AUTOCOMPLETE_OPTIONS.map(option => ({
