+++ demo/src/server.js
@@ -12,8 +12,7 @@ dotenv.config();
 // Setup app
 const app = express();
 app.use(express.urlencoded({ extended: false }));
-app.use(express.static('static'));
-app.use(express.static('dist/demo'));
+app.use(express.static('dist'));
 
 app.get('/', async (req, res, next) => {
   if (req.method.toLowerCase() === 'get' && req.headers['Content-Type'] !== 'application/json') {
