+++ package.json
@@ -18,21 +18,14 @@
   "main": "dist/index.js",
   "license": "MIT",
   "scripts": {
-    "clean": "rimraf dist",
-    "transpile": "babel src --out-dir dist --copy-files",
-    "build": "babel ./src --out-dir dist/demo/src --copy-files && babel ./demo --out-dir dist/demo/demo && webpack",
-    "start": "node ./dist/demo/demo/server.js",
-    "lint:js": "eslint src/**/*.js && eslint src/**/*.jsx && eslint demo/**/*.js && eslint demo/**/*.jsx",
-    "lint:css": "stylelint \"./src/**/*.css\"",
-    "lint:fix": "eslint --fix src/**/*.js && eslint --fix src/**/*.jsx && eslint --fix demo/**/*.js && eslint --fix demo/**/*.jsx",
-    "lint": "npm run lint:js && npm run lint:css",
+    "build:demo": "cd demo && npm run build && cd ..",
+    "build": "BABEL_ENV=production babel src --out-dir dist --copy-files && npm run build:demo",
+    "start": "cd demo && node ./dist/demo/app/server.js",
     "pretest": "jest --clearCache",
     "test": "jest",
     "ci:test": "jest --updateSnapshot --coverage",
-    "publish:component": "NODE_ENV=production npm run clean && npm run transpile && npm publish --access public"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.6.3"
+    "publish:component": "BABEL_ENV=production npm run build && npm publish --access public",
+    "postinstall": "cd demo && npm install && cd .."
   },
   "devDependencies": {
     "@babel/cli": "^7.6.4",
@@ -44,47 +37,24 @@
     "@babel/plugin-transform-runtime": "^7.6.2",
     "@babel/preset-env": "^7.6.3",
     "@babel/preset-react": "^7.6.3",
-    "babel-eslint": "^10.0.3",
+    "@babel/runtime": "^7.6.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-css-modules-transform": "^1.6.2",
     "babel-plugin-macros": "^2.6.1",
-    "cross-env": "^6.0.3",
-    "css-loader": "^3.2.0",
-    "dotenv": "^8.0.0",
+    "babel-preset-minify": "^0.5.1",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.1",
-    "eslint": "^6.6.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-standard": "^4.0.1",
-    "express": "^4.17.1",
-    "fetch-everywhere": "^1.0.5",
     "husky": "^3.0.8",
     "jest": "24.9.0",
     "jest-enzyme": "^7.1.1",
     "jsdom": "^15.1.1",
-    "mini-css-extract-plugin": "^0.8.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-cssnext": "^3.1.0",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
     "prop-types": "15.7.2",
     "react": "16.11.0",
     "react-dom": "16.11.0",
     "react-router-dom": "^5.0.1",
-    "react-test-renderer": "^16.11.0",
-    "rimraf": "^3.0.0",
-    "style-loader": "^1.0.0",
-    "stylelint": "^11.1.1",
-    "stylelint-config-recommended": "^3.0.0",
-    "terser-webpack-plugin": "^2.1.3",
-    "typescript": "3.3.3",
-    "webpack": "^4.41.1",
-    "webpack-cli": "^3.3.9"
+    "react-test-renderer": "^16.11.0"
   },
   "peerDependencies": {
     "react": "^16.8.0",
